[chore] Remove telemetryInitializer from CollectorSettings
This will allow to move deprecated Collector to otelcol.
